Add files via upload
As an intern, I've crafted a Java program named "GradeCalculator" designed to assist in calculating the grades of students based on their marks in various subjects. This interactive application prompts users to input the number of subjects they are studying and then enter the marks they've obtained in each subject (out of 100).